Improve ability to stop OrientationSensor monitoring for slower devices
On slower devices at least, starting to monitor the Magnetometer or the OrientationSensor locks the user out of using the app because it spams the app with new readings to manage. Limiting the amount of readings that awaits handling to one at a time reduces the time to wait between a click on the button to stop it and the moment that click gets processed.