Make F# and VB version of OpenSilver.Showcase target .NET8
Note: this is because them targetting netstandard2.0 somehow causes errors when trying to build the C# version (when they are loaded in the solution) - The error says to check that the project properly targets .NET8.

I did not test whether the F# and VB projects build - I think they had issues beforehand anyway.